Updated FindCustomer Logic

+++ spring-boot-web-ibs/src/main/java/com/wellsfargo/batch7/sbwibs/repo/CustomerRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.wellsfargo.batch7.sbwibs.entity.CustomerEntity;
 
 @Repository
-public interface CustomerRepo extends JpaRepository<CustomerEntity, Integer> {
+public interface CustomerRepo extends JpaRepository<CustomerEntity, Long> {
 
 	boolean existsByEmailId(String emailId);
 	boolean existsByDateOfBirth(String dateOfBirth);
